Updated Pages
Getting ready to deploy some sort of product. The major issue in this update is some jinja error where it is finding an unexpected character \\. I'm not sure why this is happening and it seems to be an issue with the LaTex syntax. I don't have this error with the unit6 page's text.